+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Feat(mobile): Add in-app survey feature. The coordinator can trigger surveys which will be shown in the app.
+
 ## [1.8.5] - 2024-02-05
 
 - Feat(webapp): Show order history